회원가입 성공 시 자동 로그인 기능 구현 & 전화번호 형식 체크 기능 구현
