+++ src/pages/Login/signUp/AskEmail.tsx
@@ -117,7 +117,7 @@ export default function AskEmail() {
 			<Button className="next" onClick={handleClick}>
 				다음
 			</Button>
-			<Button className="option" onClick={() => navigate('/signUp/email')}>
+			<Button className="option" onClick={() => navigate('/signUp/phone')}>
 				휴대폰 번호로 가입
 			</Button>
 			<Button className="already" onClick={() => navigate('/')}>
